add the authorization middleware in the role controller

+++ SelXPressApi/SelXPressApi/Controllers/RoleController.cs
@@ -1,8 +1,10 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using System.Net;
+using Microsoft.AspNetCore.Mvc;
 using SelXPressApi.DocumentationErrorTemplate;
 using SelXPressApi.DTO.RoleDTO;
 using SelXPressApi.Exceptions;
 using SelXPressApi.Interfaces;
+using SelXPressApi.Middleware;
 using SelXPressApi.Models;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
@@ -14,25 +16,38 @@ namespace SelXPressApi.Controllers
 	public class RoleController : ControllerBase
 	{
 		private readonly IRoleRepository _roleRepository;
+		private readonly IAuthorizationMiddleware _authorizationMiddleware;
 
-		public RoleController(IRoleRepository roleRepository)
+		public RoleController(IRoleRepository roleRepository, IAuthorizationMiddleware authorizationMiddleware)
 		{
 			_roleRepository = roleRepository;
+			_authorizationMiddleware = authorizationMiddleware;
 		}
+		
 		/// <summary>
-		/// GET: api/<RoleController>
-		/// Get all roles
+		/// Method to get all the roles of the database
 		/// </summary>
-		/// <returns>Return an Array with all roles </returns>
+		/// <returns></returns>
+		/// <exception cref="ForbiddenRequestException"></exception>
+		/// <exception cref="BadRequestException"></exception>
+		/// <exception cref="NotFoundException"></exception>
 		[HttpGet]
 		[ProducesResponseType(200, Type = typeof(List<Role>))]
 		[ProducesResponseType(400, Type = typeof(BadRequestErrorTemplate))]
+		[ProducesResponseType(401, Type = typeof(UnauthorizedErrorTemplate))]
+		[ProducesResponseType(403, Type = typeof(ForbiddenErrorTemplate))]
 		[ProducesResponseType(404, Type = typeof(NotFoundErrorTemplate))]
 		[ProducesResponseType(500, Type = typeof(InternalServerErrorTemplate))]
 		public async Task<IActionResult> GetRoles()
 		{
+			await _authorizationMiddleware.CheckIfTokenExists(HttpContext);
+			
+			if (!await _authorizationMiddleware.CheckRoleIfAdmin(HttpContext))
+				throw new ForbiddenRequestException("You are not authorized to do this operation", "todo");
+			
 			if (!ModelState.IsValid)
 				throw new BadRequestException("The model is wrong, a bad request occured", "RLE-1101");
+			
 			var roles = await _roleRepository.GetAllRoles();
 
 			if (roles.Count == 0)
@@ -42,18 +57,27 @@ namespace SelXPressApi.Controllers
 		}
 
 		/// <summary>
-		/// GET api/<RoleController>/5
-		/// Get a role by id
+		/// Return the role based on the parameter id
 		/// </summary>
 		/// <param name="id"></param>
-		/// <returns>Return a list of user with a specific role</returns>
+		/// <returns></returns>
+		/// <exception cref="ForbiddenRequestException"></exception>
+		/// <exception cref="NotFoundException"></exception>
+		/// <exception cref="BadRequestException"></exception>
 		[HttpGet("{id}")]
 		[ProducesResponseType(200, Type = typeof(Role))]
 		[ProducesResponseType(400, Type = typeof(BadRequestErrorTemplate))]
+		[ProducesResponseType(401, Type = typeof(UnauthorizedErrorTemplate))]
+		[ProducesResponseType(403, Type = typeof(ForbiddenErrorTemplate))]
 		[ProducesResponseType(404, Type = typeof(NotFoundErrorTemplate))]
 		[ProducesResponseType(500, Type = typeof(InternalServerErrorTemplate))]
 		public async Task<IActionResult> GetRole(int id)
 		{
+			await _authorizationMiddleware.CheckIfTokenExists(HttpContext);
+
+			if (!await _authorizationMiddleware.CheckRoleIfAdmin(HttpContext))
+				throw new ForbiddenRequestException("You are not authorized to do this operation", "todo");
+			
 			if (!await _roleRepository.RoleExists(id))
 				throw new NotFoundException("The role with the id : " + id + " doesn't exist", "RLE-1402");
 			if (!ModelState.IsValid)
@@ -63,38 +87,56 @@ namespace SelXPressApi.Controllers
 		}
 
 		/// <summary>
-		/// POST api/<RoleController>
-		/// Create a new role
+		/// Method to create a role
 		/// </summary>
-		/// <param name="value"></param>
+		/// <param name="role"></param>
+		/// <returns></returns>
+		/// <exception cref="ForbiddenRequestException"></exception>
+		/// <exception cref="BadRequestException"></exception>
+		/// <exception cref="Exception"></exception>
 		[HttpPost]
 		[ProducesResponseType(201)]
 		[ProducesResponseType(400 , Type = typeof(BadRequestErrorTemplate))]
+		[ProducesResponseType(401, Type = typeof(UnauthorizedErrorTemplate))]
+		[ProducesResponseType(403, Type = typeof(ForbiddenErrorTemplate))]
 		[ProducesResponseType(500, Type = typeof(InternalServerErrorTemplate))]
 		public async Task<IActionResult> CreateRole([FromBody] CreateRoleDTO role)
 		{
+			await _authorizationMiddleware.CheckIfTokenExists(HttpContext);
+			
+			if (!await _authorizationMiddleware.CheckRoleIfAdmin(HttpContext))
+				throw new ForbiddenRequestException("You are not authorized to do this operation", "todo");
+			
 			if (role == null || role.RoleName == null)
 				throw new BadRequestException("There are missing fields, please try again with some data", "RLE-1102");
-			if (await _roleRepository.CreateRole(role))
-			{
-				return StatusCode(201);
-			}
-			throw new Exception("An error occured while the creating of the role");
+			
+			await _roleRepository.CreateRole(role);
+			return StatusCode(201);
 		}
 
 		/// <summary>
-		/// PUT api/<RoleController>/5
-		/// Modify a role
+		/// Method to update the role based on an id
 		/// </summary>
 		/// <param name="id"></param>
-		/// <param name="value"></param>
+		/// <param name="updateRole"></param>
+		/// <returns></returns>
+		/// <exception cref="ForbiddenRequestException"></exception>
+		/// <exception cref="BadRequestException"></exception>
+		/// <exception cref="NotFoundException"></exception>
 		[HttpPut("{id}")]
 		[ProducesResponseType(200)]
 		[ProducesResponseType(400, Type = typeof(BadRequestErrorTemplate))]
+		[ProducesResponseType(401, Type = typeof(UnauthorizedErrorTemplate))]
+		[ProducesResponseType(403, Type = typeof(ForbiddenErrorTemplate))]
 		[ProducesResponseType(404, Type = typeof(NotFoundErrorTemplate))]
 		[ProducesResponseType(500, Type = typeof(InternalServerErrorTemplate))]
 		public async Task<IActionResult> UpdateRole(int id, [FromBody] UpdateRoleDTO updateRole)
 		{
+			await _authorizationMiddleware.CheckIfTokenExists(HttpContext);
+			
+			if (!await _authorizationMiddleware.CheckRoleIfAdmin(HttpContext))
+				throw new ForbiddenRequestException("You are not authorized to do this operation", "todo");
+			
 			if (!ModelState.IsValid)
 				throw new BadRequestException("The model is wrong, a bad request occured", "RLE-1101");
 			if (updateRole == null)
@@ -106,21 +148,33 @@ namespace SelXPressApi.Controllers
 		}
 
 		/// <summary>
-		/// DELETE api/<RoleController>/5
-		/// Delete a role
+		/// Method to delete a role
 		/// </summary>
 		/// <param name="id"></param>
+		/// <returns></returns>
+		/// <exception cref="ForbiddenRequestException"></exception>
+		/// <exception cref="NotFoundException"></exception>
+		/// <exception cref="BadRequestException"></exception>
 		[HttpDelete("{id}")]
 		[ProducesResponseType(200)]
 		[ProducesResponseType(400, Type = typeof(BadRequestErrorTemplate))]
+		[ProducesResponseType(401, Type = typeof(UnauthorizedErrorTemplate))]
+		[ProducesResponseType(403, Type = typeof(ForbiddenErrorTemplate))]
 		[ProducesResponseType(404, Type = typeof(NotFoundErrorTemplate))]
 		[ProducesResponseType(500, Type = typeof(InternalServerErrorTemplate))]
 		public async Task<IActionResult> DeleteRole(int id)
 		{
+			await _authorizationMiddleware.CheckIfTokenExists(HttpContext);
+			
+			if (!await _authorizationMiddleware.CheckRoleIfAdmin(HttpContext))
+				throw new ForbiddenRequestException("You are not authorized to do this operation", "todo");
+			
 			if (!await _roleRepository.RoleExists(id))
 				throw new NotFoundException("The role with the id : " + id + " doesn't exist", "RLE-1402");
+			
 			if (!ModelState.IsValid)
 				throw new BadRequestException("The model is wrong, a bad request occured","RLE-1101");
+			
 			await _roleRepository.DeleteRole(id);
 			return Ok();
 		}
